feat(tracker): complete Phase 02 historical bars service
- Mark Phase 02 complete in impl.md and tick all phase-02 checkboxes
- Fix clippy: silence too_many_arguments on insert_bar test helper, drop redundant i64 casts
- Apply rustfmt to mod.rs and tests.rs
- Document HistoricalDataService, HistoricalRateLimiter, tracker.rs commands, and bars_cache write path in CLAUDE.md
- Append Phase 02 entry to schema-decisions.md (no schema change; PK is sufficient)

+++ src-tauri/src/services/historical_data_service/tests.rs
@@ -106,6 +106,7 @@ fn rate_limiter() -> Arc<HistoricalRateLimiter> {
     Arc::new(HistoricalRateLimiter::new(60))
 }
 
+#[allow(clippy::too_many_arguments)]
 async fn insert_bar(
     db: &Db,
     symbol: &str,
@@ -245,7 +246,7 @@ async fn cache_miss_fetches_and_writes_through() {
                 high: 201.0 + i as f64,
                 low: 199.0 + i as f64,
                 close: 200.5 + i as f64,
-                volume: 5_000 + i as i64,
+                volume: 5_000 + i,
                 wap: 200.25 + i as f64,
                 count: 100,
             }
@@ -488,7 +489,11 @@ async fn rate_limiter_invoked_per_request() {
             .await
             .expect("fetch ok");
     }
-    assert_eq!(rl.acquire_count().await, 3, "three fetches → three acquires");
+    assert_eq!(
+        rl.acquire_count().await,
+        3,
+        "three fetches → three acquires"
+    );
 
     // Pre-populate cache for DDD so the next call is a cache hit.
     insert_bar(
@@ -534,7 +539,7 @@ async fn service_dedups_in_flight_requests_for_same_key() {
                 high: 101.0 + i as f64,
                 low: 99.0 + i as f64,
                 close: 100.5 + i as f64,
-                volume: 1_000 + i as i64,
+                volume: 1_000 + i,
                 wap: 100.25 + i as f64,
                 count: 1,
             }
